feat: ResourceLoader request triggers action ONLY
Added new functions to `loadResourceRequest` and `updateResourceRequest`
which dispatch the actions used in `loadResource` and `updateResource`
and returns the redux-saga-thunk promise returned.

This makes it possible to request the resource without making the
component change status states. Since it returns the redux-saga-thunk
promise success and error handling is easily accessible.
